Move some element to base class.
