+++ src/lmic/radio_sx1272.cpp
@@ -560,14 +560,6 @@ void Radio::handle_end_tx() const {
   opmode(OPMODE_SLEEP);
 }
 
-int16_t Radio::get_last_packet_rssi() const {
-  // see documentation for -139
-  // do not handle snr > 0
-  return -139 + last_packet_rssi_reg;
-}
-
-int8_t Radio::get_last_packet_snr_x4() const { return last_packet_snr_reg; }
-
 void Radio::rst() const {
   DisableIRQsGard irqguard;
   // put radio to sleep
