+++ src/lmic/radio_sx1276.cpp
@@ -393,16 +393,6 @@ void RadioSx1276::handle_end_tx() const {
   opmode(OPMODE_SLEEP);
 }
 
-int16_t RadioSx1276::get_last_packet_rssi() const {
-  // see documentation for -139
-  // do not handle snr > 0
-  return -139 + last_packet_rssi_reg;
-}
-
-int8_t RadioSx1276::get_last_packet_snr_x4() const {
-  return last_packet_snr_reg;
-}
-
 void RadioSx1276::rst() const {
   DisableIRQsGard irqguard;
   // put radio to sleep
