+++ src/lmic/radio_sx1276.h
@@ -35,12 +35,8 @@ public:
 
   bool io_check() const override;
   uint8_t rssi() const override;
-  int16_t get_last_packet_rssi() const override;
-  int8_t get_last_packet_snr_x4() const override;
 
 private:
-  int8_t last_packet_snr_reg = 0;
-  uint8_t last_packet_rssi_reg = 0;
   HalIo hal;
 
   void opmode(uint8_t mode) const;
